[refactor:move] xtask: main.rs(1,029行)を機能別7モジュールへ分割
main.rs を dispatch + doctor + print_usage の92行に縮小し、
exec(実行基盤)/rust/desktop/cn/scenario/release/oversized の
7モジュールへ関数を移動。既存 unit テスト7本は対象モジュール内の
#[cfg(test)] へ分散移動。

挙動変更なし: 18コマンド名・usage 出力・oversized 出力形式・
RUST_MIN_STACK 注入・e2e-smoke の 64MiB スタックスレッド・
cn env 2系統は全て不変。可視性の pub(crate) 付与と
`use crate::*` glob(main.rs の pub(crate) use 再輸出経由)のみ。

これにより xtask 自身が oversized リストから外れ(13→12件)、
後続の ratchet 化(WP-S2 T2)の実装先が確保される。

検証: cargo test -p xtask (7/7)、clippy --all-targets -D warnings、
fmt --check、oversized-files/doctor/不正引数 の動作確認。

Refs: .claude/plans/2026-07-03-refactor-s2-quality-gates.md (T1 / 付録A)
